requirements.txt
wikipedia==1.4.0
google-genai
python-dotenv
litellm